- Fixed SavedGame status update in HomeViewModel - Fixed GetLastGameUseCase

+++ features/home/domain/src/main/kotlin/io/github/japskiddin/sudoku/feature/home/domain/usecase/DeleteSavedGameUseCase.kt
@@ -2,6 +2,7 @@ package io.github.japskiddin.sudoku.feature.home.domain.usecase
 
 import io.github.japskiddin.sudoku.core.domain.SavedGameRepository
 import io.github.japskiddin.sudoku.core.model.SavedGame
+import io.github.japskiddin.sudoku.core.model.isCurrent
 import javax.inject.Inject
 
 public class DeleteSavedGameUseCase
@@ -10,6 +11,8 @@ constructor(
     private val savedGameRepository: SavedGameRepository
 ) {
     public suspend operator fun invoke(savedGame: SavedGame) {
-        savedGameRepository.delete(savedGame)
+        if (savedGame.isCurrent()) {
+            savedGameRepository.delete(savedGame)
+        }
     }
 }
